LayoutTests/fast/xmlhttprequest/xmlhttprequest-no-file-access-expected.txt

CONSOLE MESSAGE: line 23: XMLHttpRequest cannot load xmlhttprequest-no-file-access-expected.txt. Cross origin requests are only supported for HTTP.

The child iframe cannot paste its textual results into this iframe because it is considered a different domain - that's the point of this test! Therefore, success is denoted by the child iframe calling notifyDone. The test will hang if something goes amiss with the access control checks.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------

We're checking we can't read an arbitrary file when we set each file:// URI to have a unique domain.

Checking that same-origin iframes work.
Doing an XHR to an existing file.
Exception: A network error occurred.
Exception: Blocked a frame with origin "null" from accessing a frame with origin "file://". Protocols, domains, and ports must match.


--------
Frame: 'f'
--------
Successful write into iframe
